gvle: add a basic model configure service
From the contextual menu of the DEVS diagram a "configure" action is
now available. This action is enabled only if the model is not
configured yet. A non configured model is a model without ports,
dynamics, observables, and conditions.
The configure action provide a submenu exibiting all the classes of
models locally installed and which have a metadata associated to.
When choosing a class the model should be configured as, the
configuration action import and set to the model observable, ports,
condition, and the dynamic as provided inside the metatdata of the
class.